requirements.txt
streamlit
langchain
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
pypdf
langchain-google-genai